Module msgs: Add functionality to read msgs. Change name var ad to item in frontend because adblock recognize. Implemented notification on dashboard. Added maps and address on detail ad Added template to store Added configuration to privacity for user

+++ apps/userProfile/serializers.py
@@ -1,13 +1,11 @@
 from rest_framework.serializers import ModelSerializer
 from rest_framework import serializers
 
-from .models import UserProfile, Phone, UserLocation, Store
+from .models import UserProfile, Phone, UserLocation, Store, REQUIRED_INFO_ADDRESS, CONFIG_PRIVACY
 
 from sorl.thumbnail import get_thumbnail
 
 from django.contrib.auth.models import User
-
-from apps.ad.models import Ad
 
 
 class StoreSerializer(ModelSerializer):
@@ -26,7 +24,6 @@ class StoreSerializer(ModelSerializer):
 
 
 class PhoneSerializer(ModelSerializer):
-
     class Meta:
         model = Phone
         fields = ('number', 'type', 'id')
@@ -58,6 +55,61 @@ class UserProfileSerializer(serializers.ModelSerializer):
         return super(UserProfileSerializer, self).update(instance, validated_data)
 
 
+class ProfileLocationSerializer(ModelSerializer):
+    """
+        Serializer to Userlocation for profile.
+    """
+    address = serializers.DictField()
+
+    class Meta:
+        model = UserLocation
+        fields = ('title', 'lat', 'lng', 'address', 'id')
+        read_only_fields = ('id', )
+        excluded = ('userProfile', 'is_address', 'radius')
+        depth = 1
+
+    def validate(self, attrs):
+        for key in REQUIRED_INFO_ADDRESS:
+            if not key in attrs['address'] or not attrs['address'][key]:
+                raise serializers.ValidationError("The param - %s - to address is required" %(key))
+
+        return super(ProfileLocationSerializer, self).validate(attrs)
+
+    def __init__(self, *args, **kwargs):
+        super(ProfileLocationSerializer, self).__init__(*args, **kwargs)
+        if not self.instance:
+            request = self.context.get('request', None)
+            self.instance = UserLocation()
+            self.instance.userProfile = request.user.profile
+            self.instance.radius = 5000
+            self.instance.is_address = True
+
+
+    def create(self, validated_data):
+        request = self.context.get('request', None)
+        if request is not None:
+            if request.user.is_authenticated():
+                validated_data['userProfile'] =  request.user.profile
+                validated_data['radius'] = 5000
+                validated_data['is_address'] = True
+        return UserLocation.objects.create(**validated_data)
+
+
+
+class ConfigPrivacitySerializer(ModelSerializer):
+    config = serializers.DictField(source='privacy_settings')
+
+    def validate(self, attrs):
+        for key, value in CONFIG_PRIVACY.items():
+            if not key in attrs['privacy_settings'] or not type(attrs['privacy_settings'][key]) is bool:
+                raise serializers.ValidationError("The param %s to config privacity is required" %(key))
+
+        return super(ConfigPrivacitySerializer, self).validate(attrs)
+
+    class Meta:
+        model = UserProfile
+        fields = ('config',)
+        depth = 1
 
 
 class ProfileSerializer(ModelSerializer):
@@ -72,6 +124,7 @@ class ProfileSerializer(ModelSerializer):
         fields = ('image', 'birth_date', 'user', 'phones', 'thumbnail_200x200', 'id')
         read_only_fields = ('id',)
 
+
     def update(self, instance, validated_data):
         user_data = validated_data.get('user', instance.user)
         request = self.context.get('request', None)
@@ -79,15 +132,17 @@ class ProfileSerializer(ModelSerializer):
         instance.user = user_serializer.update(instance=request.user, validated_data=user_data)
 
         phones_data = validated_data.get('phones', [])
-        Phone.objects.filter(userProfile=instance).delete()
-        for phone in phones_data:
-            p = Phone()
-            p.userProfile = instance
-            p.number= int(phone.get('number', 0))
-            p.type = phone.get('type', 'TEL')
-            p.save()
+        if phones_data:
+            Phone.objects.filter(userProfile=instance).delete()
+            for phone in phones_data:
+                p = Phone()
+                p.userProfile = instance
+                p.number= int(phone.get('number', 0))
+                p.type = phone.get('type', 'TEL')
+                p.save()
 
         instance.birth_date = validated_data.get('birth_date', instance.birth_date)
+
         instance.save()
 
         return instance
